[INIT/#1] ktlint 추가
