+++ .editorconfig
@@ -1,0 +1,7 @@
+root = true
+
+[*]
+max_line_length = off
+
+[*.{kt,kts}]
+ktlint_function_naming_ignore_when_annotated_with=Composable
