+++ app/src/androidTest/java/org/sopt/linkareer/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package org.sopt.linkareer_android
+package org.sopt.linkareer
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
