+++ app/src/test/java/org/sopt/linkareer/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package org.sopt.linkareer_android
+package org.sopt.linkareer
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
